added errors handling, renamed fields
