+++ client/mocks/builders_test.go
@@ -996,55 +996,6 @@ func buildRdsDBSubnetGroups(t *testing.T, ctrl *gomock.Controller) client.Servic
 	}
 }
 
-func buildIamGroups(t *testing.T, ctrl *gomock.Controller) client.Services {
-	m := mocks.NewMockIamClient(ctrl)
-	g := iamTypes.Group{}
-	err := faker.FakeData(&g)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	p := iamTypes.AttachedPolicy{}
-	err = faker.FakeData(&p)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	m.EXPECT().ListGroups(gomock.Any(), gomock.Any(), gomock.Any()).Return(
-		&iam.ListGroupsOutput{
-			Groups: []iamTypes.Group{g},
-		}, nil)
-	m.EXPECT().ListAttachedGroupPolicies(gomock.Any(), gomock.Any(), gomock.Any()).Return(
-		&iam.ListAttachedGroupPoliciesOutput{
-			AttachedPolicies: []iamTypes.AttachedPolicy{p},
-		}, nil)
-
-	//list policies
-	var l []string
-	err = faker.FakeData(&l)
-	if err != nil {
-		t.Fatal(err)
-	}
-	m.EXPECT().ListGroupPolicies(gomock.Any(), gomock.Any(), gomock.Any()).Return(
-		&iam.ListGroupPoliciesOutput{
-			PolicyNames: l,
-		}, nil)
-
-	//get policy
-	gp := iam.GetGroupPolicyOutput{}
-	err = faker.FakeData(&gp)
-	if err != nil {
-		t.Fatal(err)
-	}
-	document := "{\"test\": {\"t1\":1}}"
-	gp.PolicyDocument = &document
-	m.EXPECT().GetGroupPolicy(gomock.Any(), gomock.Any(), gomock.Any()).Return(
-		&gp, nil)
-	return client.Services{
-		IAM: m,
-	}
-}
-
 func buildIamPolicies(t *testing.T, ctrl *gomock.Controller) client.Services {
 	m := mocks.NewMockIamClient(ctrl)
 	g := iamTypes.ManagedPolicyDetail{}
