+++ resources/iam_groups.go
@@ -32,8 +32,8 @@ func IamGroups() *schema.Table {
 				Resolver: client.ResolveAWSRegion,
 			},
 			{
-				Name:     "policies",
-				Type:     schema.TypeStringArray,
+				Name:     "attached_policies",
+				Type:     schema.TypeJSON,
 				Resolver: resolveIamGroupPolicies,
 			},
 			{
@@ -85,12 +85,12 @@ func fetchIamGroups(ctx context.Context, meta schema.ClientMeta, parent *schema.
 func resolveIamGroupPolicies(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
 	r := resource.Item.(types.Group)
 	svc := meta.(*client.Client).Services().IAM
-	config := iam.ListGroupPoliciesInput{
+	config := iam.ListAttachedGroupPoliciesInput{
 		GroupName: r.GroupName,
 	}
-	var policies []string
+	policies := make(map[string]interface{})
 	for {
-		response, err := svc.ListGroupPolicies(ctx, &config)
+		response, err := svc.ListAttachedGroupPolicies(ctx, &config)
 		if err != nil {
 			var ae smithy.APIError
 			if errors.As(err, &ae) && ae.ErrorCode() == "NoSuchEntity" {
@@ -98,7 +98,10 @@ func resolveIamGroupPolicies(ctx context.Context, meta schema.ClientMeta, resour
 			}
 			return err
 		}
-		policies = append(policies, response.PolicyNames...)
+		for _, p := range response.AttachedPolicies {
+			policies[*p.PolicyName] = p.PolicyArn
+		}
+
 		if response.Marker == nil {
 			break
 		}
